+++ metrics/metrics-dropwizard/src/main/java/org/apache/james/metrics/dropwizard/DropWizardMetricFactory.java
@@ -81,5 +81,6 @@ public class DropWizardMetricFactory implements MetricFactory, Startable {
     @PreDestroy
     public void stop() {
         jmxReporter.stop();
+        metricRegistry.removeMatching((name, metric) -> true);
     }
 }