+++ server/container/guice/protocols/webadmin/src/main/java/org/apache/james/modules/server/WebAdminServerModule.java
@@ -78,6 +78,7 @@ public class WebAdminServerModule extends AbstractModule {
     protected void configure() {
         install(new TaskRoutesModule());
         install(new HealthCheckRoutesModule());
+        install(new MetricsRoutesModule());
 
         bind(JsonTransformer.class).in(Scopes.SINGLETON);
         bind(WebAdminServer.class).in(Scopes.SINGLETON);